requirements.txt
pyaudiowpatch
sounddevice
numpy
pystray
Pillow
pyinstaller
